Organize with one js file per component.
